+++ docs/changelog.md
@@ -2,12 +2,13 @@
 
 ## Unreleased
 
+- Fix error on use of `AbstractAsyncContextManager` (#350)
 - Check `with` and `async with` statements (#344)
 - Improve type compatibility between generics and literals (#346)
 - Infer signatures for method wrapper objects (bound methods
   of builtin types) (#345)
 - Allow storing type narrowing constraints in variables (#343)
-- The first argument to `__new__` and `__init_subclass`
+- The first argument to `__new__` and `__init_subclass__`
   does not need to be `self` (#342)
 - Drop dependencies on `attrs` and `mypy_extensions` (#341)
 - Correct location of error for incompatible parameter (#339)